+++ microservicio-inventario/src/main/java/com/reposicion/inventario/service/ProductoService.java
@@ -1,12 +1,13 @@
 package com.reposicion.inventario.service;
 
 import com.reposicion.inventario.dto.ProductoDTO;
+import com.reposicion.inventario.excepciones.ExceptionConflict;
 import com.reposicion.inventario.model.Categoria;
 import com.reposicion.inventario.model.Producto;
 import com.reposicion.inventario.repository.CategoriaRepository;
 import com.reposicion.inventario.repository.ProductoRepository;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -14,16 +15,18 @@ import java.util.List;
 
 @Slf4j
 @Service
+@RequiredArgsConstructor
 public class ProductoService {
 
-    @Autowired
-    private ProductoRepository productoRepository;
-    @Autowired
-    private CategoriaRepository categoriaRepository;
+    private final ProductoRepository productoRepository;
+    private final CategoriaRepository categoriaRepository;
 
     // Crear producto
     public Producto crearProducto(ProductoDTO productoDTO) {
         log.info("Iniciando creacion de un nuevo producto con codigo de barra {}", productoDTO.getCodigoBarra());
+        if(productoRepository.existsByCodigoBarra(productoDTO.getCodigoBarra())){
+            throw new ExceptionConflict("El codigo de barra " + productoDTO.getCodigoBarra() + " Ya existe en otro producto.");
+        }
         Producto productoNuevo = new Producto();
         productoNuevo.setCodigoBarra(productoDTO.getCodigoBarra());
         productoNuevo.setNombreProducto(productoDTO.getNombreProducto());
@@ -47,22 +50,27 @@ public class ProductoService {
     // Listar un producto en especifico
     public Producto listarUnSoloProducto(Long id){
         log.debug("Iniciando busqueda de producto con ID {}", id);
-        return productoRepository.findById(id).orElseThrow(() -> new RuntimeException("Producto no encontrado" + id));
+        return productoRepository.findById(id).orElseThrow(() -> new RuntimeException("Producto no encontrado con ID: " + id));
     }
 
     // Actualizar producto
-    public Producto actualizarProducto(Long id, ProductoDTO productoActualizado) {
+    public Producto actualizarProducto(Long id, ProductoDTO productoDTO) {
         log.info("Iniciando actualizacion del producto con ID: {}", id);
         Producto productoExistente = productoRepository.findById(id).orElseThrow(() -> {log.error("Error al actualizar producto, ID {} no encontrado", id);
-            return new RuntimeException("Producto no encontrado" + id);
+            return new RuntimeException("Producto no encontrado con ID: " + id);
         });
-        productoExistente.setCodigoBarra(productoActualizado.getCodigoBarra());
-        productoExistente.setNombreProducto(productoActualizado.getNombreProducto());
-        productoExistente.setStock(productoActualizado.getStock());
-        productoExistente.setDescripcionProducto(productoActualizado.getDescripcionProducto());
-        productoExistente.setFechaEntradaProducto(productoActualizado.getFechaEntradaProducto());
-        if (productoActualizado.getCategoriaId() != null) {
-            Categoria categoria = categoriaRepository.findById(productoActualizado.getCategoriaId()).orElseThrow(() -> new RuntimeException("Categoria no encontrada con ID: " + productoActualizado.getCategoriaId()));
+        if (productoRepository.existsByCodigoBarra(productoDTO.getCodigoBarra())
+            && !productoExistente.getCodigoBarra().equals(productoDTO.getCodigoBarra())){
+            throw new ExceptionConflict("El codigo de barra " + productoDTO.getCodigoBarra() + " Ya existe en otro producto.");
+        }
+        productoExistente.setCodigoBarra(productoDTO.getCodigoBarra());
+        productoExistente.setNombreProducto(productoDTO.getNombreProducto());
+        productoExistente.setStock(productoDTO.getStock());
+        productoExistente.setDescripcionProducto(productoDTO.getDescripcionProducto());
+        productoExistente.setFechaEntradaProducto(productoDTO.getFechaEntradaProducto());
+        if (productoDTO.getCategoriaId() != null) {
+            Categoria categoria = categoriaRepository.findById(productoDTO.getCategoriaId()).orElseThrow(() ->
+                    new RuntimeException("Categoria no encontrada con ID: " + productoDTO.getCategoriaId()));
             productoExistente.setCategoria(categoria);
         }
         log.info("Producto actualizado exitosamente en la base de datos");
@@ -72,6 +80,9 @@ public class ProductoService {
     // Eliminar producto
     public void eliminarProducto(Long id){
         log.info("Iniciando eliminacion de producto con ID {}", id);
+        if(!productoRepository.existsById(id)){
+            throw new RuntimeException("Producto no encontrado con ID: " + id);
+        }
         productoRepository.deleteById(id);
     }
 }
